+++ bin/main.ml
@@ -4,7 +4,7 @@ open Ppx_yojson_conv_lib.Yojson_conv.Primitives
 
 let github_repo = "https://github.com/mazeppa-dev/mazeppa"
 
-let version = "0.2.0"
+let version = "0.3.0"
 
 let tab = "    "