+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## unreleased
 
+## 0.3.0 - 2024-08-16
+
 ### Added
 
  - The `--print-gc-stats` flag for `run` and `eval` to observe total memory usage.